+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+### Added
+  - PET: new module for PET images
+  - PET: DICOM file format
+
 ### Changed
   - baseModule: acquisition time is now stored in `_acqTime` attribute,
 so it can be directly set by `setAcqTime(datetime)` function. The value